split primitives
