+++ Cargo.toml
@@ -6,6 +6,7 @@ members = [
     "chain-signatures/keys",
     "chain-signatures/node",
     "chain-signatures/primitives",
+    "chain-signatures/signet-primitives",
     "integration-tests", 
     "chain-signatures/chain-integration-core", 
     "chain-signatures/chain-canton", 
@@ -92,6 +93,7 @@ mpc-chain-near = { path = "./chain-signatures/chain-near" }
 mpc-keys = { path = "./chain-signatures/keys" }
 mpc-node = { path = "./chain-signatures/node" }
 mpc-primitives = { path = "./chain-signatures/primitives" }
+signet-primitives = { path = "./chain-signatures/signet-primitives" }
 signet-program = { path = "./chain-signatures/contract-sol" }
 
 [patch.crates-io]
