+++ chain-signatures/contract/Cargo.toml
@@ -11,6 +11,7 @@ borsh.workspace = true
 k256.workspace = true
 mpc-crypto.workspace = true
 mpc-primitives.workspace = true
+signet-primitives.workspace = true
 near-sdk.workspace = true
 serde.workspace = true
 serde_json.workspace = true
