+++ chain-signatures/node/src/respond_bidirectional.rs
@@ -3,8 +3,8 @@ use alloy::primitives::Bytes;
 use k256::Scalar;
 use mpc_crypto::ScalarExt;
 use mpc_primitives::{
-    BidirectionalTx, Chain, IndexedSignRequest, RespondBidirectionalSerializedOutput,
-    RespondBidirectionalTx, SerDeserFormat, SignArgs, SignId,
+    BidirectionalTx, Chain, ChainConfig as _, IndexedSignRequest,
+    RespondBidirectionalSerializedOutput, RespondBidirectionalTx, SerDeserFormat, SignArgs, SignId,
 };
 
 const MAGIC_ERROR_PREFIX: [u8; 4] = [0xde, 0xad, 0xbe, 0xef];
