+++ chain-signatures/primitives/src/bidirectional.rs
@@ -1,9 +1,4 @@
-use crate::{Chain, Signature};
-
-// Should wrap B256 from Alloy, currently adding Alloy as a dependency pulls `alloy-sol-macro-input`, which requires Rust 1.85+
-// TODO: Use B256 from Alloy once we can bump the minimum Rust version to 1.85+
-#[derive(Debug, Clone, PartialEq, Eq, Hash, serde::Serialize, serde::Deserialize, Copy)]
-pub struct BidirectionalTxId(#[serde(with = "serde_bytes")] pub [u8; 32]);
+use crate::{BidirectionalTxId, Chain, Signature};
 
 #[derive(Debug, Clone, PartialEq, Eq, Hash, serde::Serialize, serde::Deserialize)]
 pub struct BidirectionalTx {
@@ -33,7 +28,7 @@ pub struct BidirectionalTx {
 #[derive(Hash, PartialEq, Eq, Clone, Debug, serde::Serialize, serde::Deserialize)]
 pub struct RespondBidirectionalTx {
     pub tx_id: BidirectionalTxId,
-    pub output: RespondBidirectionalSerializedOutput,
+    pub output: crate::RespondBidirectionalSerializedOutput,
     /// Opaque per-chain context blob. The producing indexer serializes its own
     /// struct (see e.g. `indexer_canton::CantonChainCtx`) into bytes; the
     /// consuming publisher deserializes it back. Backlog and protocol layers
@@ -41,8 +36,6 @@ pub struct RespondBidirectionalTx {
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub chain_ctx: Option<Vec<u8>>,
 }
-
-pub type RespondBidirectionalSerializedOutput = Vec<u8>;
 
 #[derive(Clone, Debug)]
 pub struct RespondBidirectionalEvent {
