+++ chain-signatures/primitives/src/lib.rs
@@ -1,3 +1,7 @@
+//! Node-internal primitives. Public interface types (used by contracts and
+//! clients) live in `signet-primitives` and are re-exported here so node code
+//! can keep a single import path.
+
 mod backlog;
 mod bidirectional;
 mod chain;
@@ -5,17 +9,13 @@ mod crypto;
 mod events;
 mod requests;
 
+pub use signet_primitives::*;
+
 pub use backlog::{Checkpoint, CheckpointDigest, ConsensusCheckpointDigest, PendingTx};
 pub use bidirectional::{
-    BidirectionalTx, BidirectionalTxId, RespondBidirectionalEvent,
-    RespondBidirectionalSerializedOutput, RespondBidirectionalTx, SignBidirectionalEvent,
+    BidirectionalTx, RespondBidirectionalEvent, RespondBidirectionalTx, SignBidirectionalEvent,
 };
-pub use chain::{Chain, ChainFromError, SerDeserFormat};
-pub use crypto::{
-    borsh_scalar, PublicKey, ScalarExt, SignArgs, SignId, Signature, MAX_SECP256K1_SCALAR,
-};
+pub use chain::{ChainConfig, SerDeserFormat};
+pub use crypto::{cbor_scalar, SignArgs};
 pub use events::{ChainEvent, ExecutionOutcome, SignatureRespondedEvent};
 pub use requests::{IndexedSignRequest, SignCommand, SignKind};
-
-pub const LATEST_MPC_KEY_VERSION: u32 = 1;
-pub const LEGACY_MPC_KEY_VERSION_0: u32 = 0;
