+++ chain-signatures/signet-primitives/Cargo.toml
@@ -1,17 +1,18 @@
 [package]
-name = "mpc-primitives"
+name = "signet-primitives"
 version.workspace = true
 edition = "2021"
 
 [dependencies]
 borsh.workspace = true
-ciborium.workspace = true
+enum-map.workspace = true
 hex.workspace = true
 k256.workspace = true
 serde.workspace = true
 serde_bytes.workspace = true
+serde_json.workspace = true
 sha3.workspace = true
-signet-primitives.workspace = true
+thiserror.workspace = true
 
 [target.'cfg(target_arch = "wasm32")'.dependencies]
 getrandom = { version = "0.2.12", features = ["custom"] }
